requirements.txt
boto3==1.16.63
botocore==1.19.63
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-JWT-Extended==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
PyJWT==2.1.0
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
